# specs/001-media-influence-detection/contracts/api.yaml
openapi: 3.1.0
info:
  title: MITDS API
  description: Media Influence Topology & Detection System REST API
  version: 1.0.0
  contact:
    name: MITDS Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.mitds.example/v1
    description: Production
...
paths:
...
  /detection/temporal-coordination:
    post:
      summary: Analyze temporal coordination
      tags: [Detection]
      operationId: analyzeTemporalCoordination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemporalAnalysisRequest'
      responses:
        '200':
          description: Coordination analysis results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporalAnalysisResponse'
        '202':
          description: Analysis queued (for large datasets)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobResponse'
  /detection/composite-score:
    post:
      summary: Calculate composite coordination score
      tags: [Detection]
      operationId: calculateCompositeScore
      description: >
        Combines temporal, funding, and infrastructure signals.
        No single signal can trigger a flag alone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompositeScoreRequest'
      responses:
        '200':
          description: Composite scores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompositeScoreResponse'
  /detection/explain/{finding_id}:
    get:
      summary: Get explanation for detection finding
      tags: [Detection]
      operationId: explainFinding
      parameters:
        - name: finding_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Explanation with evidence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingExplanation'

  # ============== REPORTS ==============
...
  /jobs/{id}:
    get:
      summary: Get async job status
      tags: [Ingestion, Detection, Reports]
      operationId: getJobStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
components:
  schemas:
...
    EntityType:
      type: string
      enum: [PERSON, ORGANIZATION, OUTLET, DOMAIN, PLATFORM_ACCOUNT, SPONSOR, VENDOR]
...
    EntitySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          $ref: '#/components/schemas/EntityType'
        name:
          type: string
...
    TemporalAnalysisRequest:
      type: object
      required: [outlet_ids, start_date, end_date]
      properties:
        outlet_ids:
          type: array
          items:
            type: string
            format: uuid
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        topic_filter:
          type: string
          description: Optional topic/keyword filter
    TemporalAnalysisResponse:
      type: object
      properties:
        finding_id:
          type: string
          format: uuid
        outlet_count:
          type: integer
        time_range:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        coordination_clusters:
          type: array
          items:
            $ref: '#/components/schemas/TemporalCluster'
        hard_negatives_filtered:
          type: integer
          description: Count of legitimate sync events excluded
    TemporalCluster:
      type: object
      properties:
        cluster_id:
          type: string
        outlets:
          type: array
          items:
            $ref: '#/components/schemas/EntitySummary'
        sync_score:
          type: number
          description: Jensen-Shannon divergence (lower = more synced)
        burst_periods:
          type: array
          items:
            type: object
            properties:
              start:
                type: string
                format: date-time
              end:
                type: string
                format: date-time
              intensity:
                type: number
    CompositeScoreRequest:
      type: object
      required: [outlet_ids]
      properties:
        outlet_ids:
          type: array
          items:
            type: string
            format: uuid
        weights:
          type: object
          properties:
            temporal:
              type: number
              default: 0.4
            funding:
              type: number
              default: 0.4
            infrastructure:
              type: number
              default: 0.2
    CompositeScoreResponse:
      type: object
      properties:
        finding_id:
          type: string
          format: uuid
        overall_score:
          type: number
          description: 0-1, higher = more coordinated
        signal_breakdown:
          type: object
          properties:
            temporal_score:
              type: number
            funding_score:
              type: number
            infrastructure_score:
              type: number
        flagged:
          type: boolean
          description: True only if multiple signals exceed thresholds
        confidence_band:
          type: object
          properties:
            lower:
              type: number
            upper:
              type: number
    FindingExplanation:
      type: object
      properties:
        finding_id:
          type: string
          format: uuid
        summary:
          type: string
          description: Non-accusatory explanation
        contributing_factors:
          type: array
          items:
            type: object
            properties:
              factor_type:
                type: string
              description:
                type: string
              weight:
                type: number
              evidence_refs:
                type: array
                items:
                  $ref: '#/components/schemas/EvidenceRef'
        confidence_statement:
          type: string
        limitations:
          type: array
          items:
            type: string

    # ============== EVIDENCE ==============
    EvidenceRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        evidence_type:
          type: string
        source_url:
          type: string
          format: uri
        archive_url:
          type: string
          format: uri
        retrieved_at:
          type: string
          format: date-time
...
    AsyncJobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        status_url:
          type: string
          format: uri
    JobStatus:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        result_url:
          type: string
          format: uri
        error:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []